pull out RelayRequest from relay_handler.js
